# docs/swagger.yaml
basePath: /api/v1
definitions:
  internal_transport_http_v1.authUserResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
  internal_transport_http_v1.createGoalRequest:
    properties:
      category_type:
        type: string
      color:
        type: string
      title:
        type: string
    type: object
...
  internal_transport_http_v1.goalData:
    properties:
      category_type:
        type: string
      color:
        type: string
      created_at:
        type: string
      id:
        type: integer
      is_archived:
        type: boolean
      title:
        type: string
    type: object
  internal_transport_http_v1.goalResponse:
    properties:
      category_type:
        type: string
      color:
        type: string
      created_at:
        type: string
      id:
        type: integer
      is_archived:
        type: boolean
      title:
        type: string
      user_id:
        type: integer
    type: object
  internal_transport_http_v1.listGoalsResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/internal_transport_http_v1.goalData'
        type: array
      user_id:
        type: integer
    type: object
...
  internal_transport_http_v1.loginUserRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  internal_transport_http_v1.refreshAccessTokenRequest:
    properties:
      refresh_token:
        type: string
    type: object
  internal_transport_http_v1.registerUserRequest:
    properties:
      confirm_password:
        type: string
      email:
        type: string
      password:
        type: string
    type: object
...
  internal_transport_http_v1.updateGoalRequest:
    properties:
      category_type:
        type: string
      color:
        type: string
      id:
        type: integer
      is_archived:
        type: boolean
      title:
        type: string
    type: object
  internal_transport_http_v1.updateGoalResponse:
    properties:
      category_type:
        type: string
      color:
        type: string
      id:
        type: integer
      is_archived:
        type: boolean
      title:
        type: string
      user_id:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
  description: API for Mile-Do, simple clone of TickTick.
  termsOfService: http://swagger.io/terms/
  title: Mile-Do API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: login to existing user account
      parameters:
      - description: Account Info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/internal_transport_http_v1.loginUserRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/internal_transport_http_v1.authUserResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: login user
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: refresh access token by refresh_token
      parameters:
      - description: Refresh token
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/internal_transport_http_v1.refreshAccessTokenRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/internal_transport_http_v1.authUserResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: refresh access token
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: create new user account
      parameters:
      - description: Account Info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/internal_transport_http_v1.registerUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_transport_http_v1.authUserResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: register new user
      tags:
      - auth
  /goals/:
    get:
      consumes:
      - application/json
      description: get list of goals
      parameters:
      - description: get goals by type
        in: query
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_http_v1.listGoalsResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: get goals
      tags:
      - goals
    patch:
      consumes:
      - application/json
      description: update existing goal
      parameters:
      - description: New Goal Info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/internal_transport_http_v1.updateGoalRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_http_v1.updateGoalResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: update goal
      tags:
      - goals
    post:
      consumes:
      - application/json
      description: create new unique goal
      parameters:
      - description: Goal Info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/internal_transport_http_v1.createGoalRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_transport_http_v1.goalResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: create new goal
      tags:
      - goals
  /goals/{id}:
    delete:
      consumes:
      - application/json
      description: delete goal by :id
      parameters:
      - description: Goal ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: goal has been removed
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: delete goal by :id
      tags:
      - goals
    get:
      consumes:
      - application/json
      description: get goal by :id
      parameters:
      - description: Goal ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_http_v1.goalResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: get goal by :id
      tags:
      - goals
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"